+++ .circleci/config.yml
@@ -209,7 +209,7 @@ jobs:
           environment: ./postman/environments/${K8S_ENV}.postman_environment.json
           additional-options: --env-var="Username=${BASIC_AUTH}"
       - newman/newman-run:
-          collection: ./postman/collections/hearing_summaries.postman_collection.json
+          collection: ./postman/collections/case_summaries.postman_collection.json
           environment: ./postman/environments/${K8S_ENV}.postman_environment.json
           additional-options: --env-var="Username=${BASIC_AUTH}"
       - newman/newman-run:
