+++ laa_court_data_api_app/models/case_summaries/defence_counsel.py
