--- laa_court_data_api_app/models/hearing_summaries/hearing_summaries_response.py
@@ -1,10 +1,0 @@
-from pydantic import BaseModel
-from laa_court_data_api_app.models.hearing_summaries.defendants import Defendants
-
-from laa_court_data_api_app.models.hearing_summaries.hearing_summary import HearingSummary
-
-
-class HearingSummariesResponse(BaseModel):
-    prosecution_case_reference: str | None
-    hearing_summaries: list[HearingSummary] | None
-    overall_defendants: list[Defendants]
