+++ test/routers/conftest.py
@@ -10,7 +10,7 @@ from laa_court_data_api_app.config.court_data_adaptor import CdaSettings
 from laa_court_data_api_app.models.hearing.external.hearing import Hearing
 from laa_court_data_api_app.models.hearing.external.hearing_result import HearingResult
 from laa_court_data_api_app.models.hearing_events.hearing_events_result import HearingEventsResult
-from laa_court_data_api_app.models.hearing_summaries.hearing_summary import HearingSummary
+from laa_court_data_api_app.models.case_summaries.hearing_summary import HearingSummary
 from laa_court_data_api_app.models.prosecution_cases.defendant_summary import DefendantSummary
 from laa_court_data_api_app.models.prosecution_cases.prosecution_cases import ProsecutionCases
 from laa_court_data_api_app.models.prosecution_cases.prosecution_cases_results import ProsecutionCasesResults
